+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - Added Discovery service
 - Added CodeStar service
 - Added Migration Hub service
+- Added Pinpoint service
 
 ## [0.28.0] - 2017-08-25